REFACTOR :: components/ 경로로 불러오는 코드 수정
