+++ apps/user/src/components/common/Icons/index.ts
@@ -1,1 +1,11 @@
-export * as default from './';
+export { default as AnswerIcon } from './Answer';
+export { default as ArrowDropDownIcon } from './ArrowDropDown';
+export { default as BamdoliroLogoIcon } from './Bamdoliro';
+export { default as BottomArrowIcon } from './BottomArrow';
+export { default as FaqIcon } from './Faq';
+export { default as GrayLogoIcon } from './GrayLogo';
+export { default as InstagramLogoIcon } from './Instagram';
+export { default as LeftArrowIcon } from './LeftArrow';
+export { default as LogoIcon } from './Logo';
+export { default as RightArrowIcon } from './RightArrow';
+export { default as TopArrowIcon } from './TopArrow';
