+++ apps/user/src/layouts/AppLayout.tsx
@@ -1,5 +1,5 @@
-import Footer from '@/components/common/Footer/Footer';
-import Header from '@/components/common/Header/Header';
+import { Footer } from '@/components/common';
+import { Header } from '@/components/common/';
 import { color } from '@maru/theme';
 import { ReactNode } from 'react';
 import styled, { CSSProp } from 'styled-components';
